Implement main requirements: 1. Add weather status and weather forecast services. Move external API querying logic to services. 2. Add geocoding service, mock logic. 3. Clean controllers, leave request validation logic in controllers. 4. Register weather and geocoding services through their interfaces as scoped services.

+++ WeatherForecasting.WebApi/Models/Response/Coordinates.cs
@@ -5,8 +5,8 @@ using System.Text.Json.Serialization;
 public class Coordinates
 {
 	[JsonPropertyName("lon")]
-	public double Lon { get; set; }
+	public decimal Lon { get; set; }
 
 	[JsonPropertyName("lat")]
-	public double Lat { get; set; }
+	public decimal Lat { get; set; }
 }
